+++ test/object-store/dictionary.cpp
@@ -79,12 +79,13 @@ TEST_CASE("dictionary") {
 
     object_store::Dictionary dict(r, obj, col);
     object_store::Dictionary links(r, obj, col_links);
-    auto results = dict.as_results();
+    auto keys_as_results = dict.get_keys();
+    auto values_as_results = dict.get_values();
     CppContext ctx(r);
 
     SECTION("get_realm()") {
         REQUIRE(dict.get_realm() == r);
-        REQUIRE(results.get_realm() == r);
+        REQUIRE(values_as_results.get_realm() == r);
     }
 
     std::vector<std::string> keys = {"a", "b", "c"};
@@ -97,9 +98,9 @@ TEST_CASE("dictionary") {
 
     SECTION("clear()") {
         REQUIRE(dict.size() == 3);
-        results.clear();
+        values_as_results.clear();
         REQUIRE(dict.size() == 0);
-        REQUIRE(results.size() == 0);
+        REQUIRE(values_as_results.size() == 0);
     }
 
     SECTION("get()") {
@@ -128,10 +129,24 @@ TEST_CASE("dictionary") {
             Dictionary::Iterator it = dict.begin() + ndx;
             REQUIRE((*it).first.get_string() == keys[i]);
             REQUIRE((*it).second.get_string() == values[i]);
-            auto element = results.get_dictionary_element(ndx);
+            auto element = values_as_results.get_dictionary_element(ndx);
             REQUIRE(element.first == keys[i]);
             REQUIRE(element.second.get_string() == values[i]);
+            std::string key = keys_as_results.get<StringData>(ndx);
+            REQUIRE(key == keys[i]);
+            Mixed m = keys_as_results.get_any(ndx);
+            REQUIRE(m.get_string() == keys[i]);
         }
+    }
+
+    SECTION("keys sorted") {
+        auto sorted = keys_as_results.sort({{"self", true}});
+        std::string key = sorted.get<StringData>(0);
+        REQUIRE(key == "a");
+        Mixed m = sorted.get_any(0);
+        REQUIRE(m.get_string() == "a");
+        m = sorted.get_any(4);
+        REQUIRE(m.is_null());
     }
 
     SECTION("handover") {
@@ -139,7 +154,7 @@ TEST_CASE("dictionary") {
 
         auto dict2 = ThreadSafeReference(dict).resolve<object_store::Dictionary>(r);
         REQUIRE(dict == dict2);
-        ThreadSafeReference ref(results);
+        ThreadSafeReference ref(values_as_results);
         auto results2 = ref.resolve<Results>(r).sort({{"self", true}});
         for (size_t i = 0; i < values.size(); ++i) {
             REQUIRE(results2.get<String>(i) == values[i]);
@@ -154,7 +169,7 @@ TEST_CASE("dictionary") {
     SECTION("notifications") {
         r->commit_transaction();
 
-        auto sorted = results.sort({{"self", true}});
+        auto sorted = values_as_results.sort({{"self", true}});
 
         size_t calls = 0;
         CollectionChangeSet change, rchange, srchange;
@@ -162,7 +177,7 @@ TEST_CASE("dictionary") {
             change = c;
             ++calls;
         });
-        auto rtoken = results.add_notification_callback([&](CollectionChangeSet c, std::exception_ptr) {
+        auto rtoken = values_as_results.add_notification_callback([&](CollectionChangeSet c, std::exception_ptr) {
             rchange = c;
             ++calls;
         });
@@ -185,7 +200,7 @@ TEST_CASE("dictionary") {
             r->commit_transaction();
 
             advance_and_notify(*r);
-            auto ndx = results.index_of(StringData("dade"));
+            auto ndx = values_as_results.index_of(StringData("dade"));
             REQUIRE_INDICES(change.insertions, ndx);
             REQUIRE_INDICES(rchange.insertions, ndx);
             // "dade" ends up at the end of the sorted list
@@ -199,7 +214,7 @@ TEST_CASE("dictionary") {
             r->commit_transaction();
 
             advance_and_notify(*r);
-            auto ndx = results.index_of(StringData("blueberry"));
+            auto ndx = values_as_results.index_of(StringData("blueberry"));
             REQUIRE_INDICES(change.insertions);
             REQUIRE_INDICES(change.modifications, ndx);
             REQUIRE_INDICES(change.deletions);
@@ -207,7 +222,7 @@ TEST_CASE("dictionary") {
 
         SECTION("remove value from list") {
             advance_and_notify(*r);
-            auto ndx = results.index_of(StringData("apple"));
+            auto ndx = values_as_results.index_of(StringData("apple"));
             r->begin_transaction();
             dict.erase("a");
             r->commit_transaction();